requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-filter==0.13.0
djangorestframework==3.3.3
gunicorn==19.4.5
Markdown==2.6.5
psycopg2==2.6.1
whitenoise==3.0
